pyproject.toml
[project]
name = "tweet"
version = "0.1.0"
description = ""
authors = [
    {name = "Nova, n"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "prettier (>=0.0.7,<0.0.8)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = {version = "^25.1.0", allow-prereleases = true}

